refactor loadbar to subcomponent
