+++ src/components/Carousel/components/Controller/components/Item/index.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import React, { useEffect, useState } from 'react'
+import React from 'react'
 
 import cn from 'classnames'
 
@@ -8,6 +8,7 @@ import { PlayIcon } from '@/assets/images'
 
 import { useCarousel } from '../../../..'
 import styles from './Item.module.scss'
+import { LoadBar } from './components/LoadBar'
 
 interface ItemProps extends React.HTMLAttributes<HTMLDivElement> {
   isActive: boolean
@@ -31,78 +32,11 @@ const Item = ({
   isVideoSlide: isVideoSlideProp,
   ...props
 }: ItemProps) => {
-  const {
-    getTimeUntilSlideFinished,
-    totalSlideTime,
-    slideTimeOverrideIndexToTimeMsMap,
-    timeUntilSlideFinishedOverrideIndexToTimeMsMap,
-    isVideoSlidePlaying,
-    isVideoSlide,
-    isLoading,
-    activeIndex,
-    emblaApi
-  } = useCarousel()
-
-  /**
-   * @dev This covers the edge case where the user leaves the tab and returns.
-   * emblaApi observes this and resets the autoplay duration when the tab regains focus.
-   * By listening to autoplay:timerset and re-rendering this component when it fires,
-   * we can be sure that we are keeping the animation time of the loader bar accurate.
-   */
-  const [itemKeyIndex, setItemKeyIndex] = useState(0)
-  useEffect(() => {
-    const timerSetListener = () => {
-      setItemKeyIndex(itemKeyIndex + 1)
-    }
-    if (activeIndex === itemIndex) {
-      emblaApi?.on('autoplay:timerset', timerSetListener)
-    }
-    return () => {
-      emblaApi?.off('autoplay:timerset', timerSetListener)
-    }
-  }, [emblaApi, activeIndex])
+  const { isLoading } = useCarousel()
 
   let loadBar = null
   if (showLoadBar && isActive) {
-    let timeUntilSlideFinishedMs = getTimeUntilSlideFinished?.()
-    if (
-      Object.hasOwn(timeUntilSlideFinishedOverrideIndexToTimeMsMap, itemIndex)
-    ) {
-      timeUntilSlideFinishedMs =
-        timeUntilSlideFinishedOverrideIndexToTimeMsMap[itemIndex].timeLeftInMs
-    }
-    let initialProgressPct = 0
-    let animationDurationMs = 5000
-    let thisItemSlideTotalTimeMs = totalSlideTime
-    if (Object.hasOwn(slideTimeOverrideIndexToTimeMsMap, itemIndex)) {
-      thisItemSlideTotalTimeMs = slideTimeOverrideIndexToTimeMsMap[itemIndex]
-    }
-    if (
-      typeof timeUntilSlideFinishedMs === 'number' &&
-      thisItemSlideTotalTimeMs
-    ) {
-      initialProgressPct =
-        ((thisItemSlideTotalTimeMs - timeUntilSlideFinishedMs) /
-          thisItemSlideTotalTimeMs) *
-        100
-      animationDurationMs = timeUntilSlideFinishedMs
-    }
-
-    loadBar = (
-      <div
-        className={cn(styles.loader, {
-          [styles.videoPlaying]: isVideoSlidePlaying[itemIndex] || !isVideoSlide
-        })}
-        // this is necessary to reset the animation timeline for the loader
-        key={`slide-${itemIndex}-progress-pct-${initialProgressPct}-anim-duration-${animationDurationMs}`}
-        style={{
-          // @ts-expect-error ts does not like css vars
-          '--carousel-item-initial-progress': `${initialProgressPct}%`,
-          '--carousel-item-animation-duration': `${animationDurationMs}ms`
-        }}
-        data-testid={`carousel-controller-item-loader-${itemIndex}`}
-      />
-    )
+    loadBar = <LoadBar itemIndex={itemIndex} className={styles.loader} />
   }
   let playIcon = null
   if (isVideoSlideProp) {
@@ -139,7 +73,6 @@ const Item = ({
         className
       )}
       onClick={buttonOnClick}
-      key={`carousel-controller-item-${itemKeyIndex}`}
       {...props}
     >
       {content}
